Move some code around (split `parse` and combine with `literal`)
