+++ macros/src/lib.rs
@@ -8,7 +8,6 @@ use proc_macro2::TokenStream;
 mod err;
 mod gen;
 mod ir;
-mod literal;
 mod parse;
 
 #[cfg(test)]
