+++ macros/src/parse/style.rs
@@ -1,0 +1,140 @@
+use proc_macro2::{Span, TokenStream};
+use crate::{
+    err::Error,
+    ir::{Style, Color},
+};
+use super::{
+    parse,
+    lit::expect_str_literal,
+};
+
+
+impl Style {
+    /// Parses the style specifiction assuming the token stream contains a
+    /// single string literal.
+    pub(crate) fn parse_from_tokens(tokens: TokenStream) -> Result<Self, Error> {
+        let (s, span) = parse(tokens, expect_str_literal)?;
+        Self::parse(&s, span)
+    }
+
+    /// Parses the style specification in `spec` (with `span`) and returns a token
+    /// stream representing an expression constructing the corresponding `ColorSpec`
+    /// value.
+    pub(super) fn parse(spec: &str, span: Span) -> Result<Self, Error> {
+        let mut out = Self::default();
+
+        let mut previous_fg_color = None;
+        let mut previous_bg_color = None;
+        for fragment in spec.split('+').map(str::trim).filter(|s| !s.is_empty()) {
+            let (fragment, is_bg) = match fragment.strip_prefix("bg:") {
+                Some(color) => (color, true),
+                None => (fragment, false),
+            };
+
+            // Parse/obtain color if a color is specified.
+            let color = match fragment {
+                "black" => Some(Color::Black),
+                "blue" => Some(Color::Blue),
+                "green" => Some(Color::Green),
+                "red" => Some(Color::Red),
+                "cyan" => Some(Color::Cyan),
+                "magenta" => Some(Color::Magenta),
+                "yellow" => Some(Color::Yellow),
+                "white" => Some(Color::White),
+
+                hex if hex.starts_with('#') => {
+                    let hex = &hex[1..];
+
+                    if hex.len() != 6 {
+                        let e = err!(
+                            span,
+                            "hex color code invalid: 6 digits expected, found {}",
+                            hex.len(),
+                        );
+                        return Err(e);
+                    }
+
+                    let digits = hex.chars()
+                        .map(|c| {
+                            c.to_digit(16).ok_or_else(|| {
+                                err!(span, "hex color code invalid: {} is not a valid hex digit", c)
+                            })
+                        })
+                        .collect::<Result<Vec<_>, _>>()?;
+
+                    let r = (digits[0] * 16 + digits[1]) as u8;
+                    let g = (digits[2] * 16 + digits[3]) as u8;
+                    let b = (digits[4] * 16 + digits[5]) as u8;
+
+                    Some(Color::Rgb(r, g, b))
+                },
+
+                // TODO: Ansi256 colors
+                _ => None,
+            };
+
+            // Check for duplicate color definitions.
+            let (previous_color, color_kind) = match is_bg {
+                true => (&mut previous_bg_color, "background"),
+                false => (&mut previous_fg_color, "foreground"),
+            };
+            match (&color, *previous_color) {
+                (Some(_), Some(old)) => {
+                    let e = err!(
+                        span,
+                        "found '{}' but the {} color was already specified as '{}'",
+                        fragment,
+                        color_kind,
+                        old,
+                    );
+                    return Err(e);
+                }
+                (Some(_), None) => *previous_color = Some(fragment),
+                _ => {}
+            }
+
+            macro_rules! set_attr {
+                ($field:ident, $value:expr) => {{
+                    if let Some(b) = out.$field {
+                        let field_s = stringify!($field);
+                        let old = if b { field_s.into() } else { format!("!{}", field_s) };
+                        let new = if $value { field_s.into() } else { format!("!{}", field_s) };
+                        let e = err!(
+                            span,
+                            "invalid style definition: found '{}', but '{}' was specified before",
+                            new,
+                            old,
+                        );
+                        return Err(e);
+                    }
+
+                    out.$field = Some($value);
+                }};
+            }
+
+            // Obtain the final token stream for method call.
+            match (is_bg, color, fragment) {
+                (false, Some(color), _) => out.fg = Some(color),
+                (true, Some(color), _) => out.bg = Some(color),
+                (true, None, other) => {
+                    return Err(err!(span, "'{}' (following 'bg:') is not a valid color", other));
+                }
+
+                (false, None, "bold") => set_attr!(bold, true),
+                (false, None, "!bold") => set_attr!(bold, false),
+                (false, None, "italic") => set_attr!(italic, true),
+                (false, None, "!italic") => set_attr!(italic, false),
+                (false, None, "underline") => set_attr!(underline, true),
+                (false, None, "!underline") => set_attr!(underline, false),
+                (false, None, "intense") => set_attr!(intense, true),
+                (false, None, "!intense") => set_attr!(intense, false),
+
+                (false, None, other) => {
+                    return Err(err!(span, "invalid style spec fragment '{}'", other));
+                }
+            }
+        }
+
+        Ok(out)
+    }
+}
